Moved Labfront Loader into appropriate structure
