+++ src/pywearable/loader/base.py
@@ -8,6 +8,14 @@ class BaseLoader:
     def get_user_ids(self) -> list:
         pass
 
+    def load_daily_summary(
+        self,
+        user_id: str,
+        start_date: Union[datetime.datetime, datetime.date, str, None] = None,
+        end_date: Union[datetime.datetime, datetime.date, str, None] = None,
+    ) -> pd.DataFrame:
+        pass
+
     def load_sleep_summary(
         self,
         user_id: str,
