+++ src/pywearable/loader/labfront/constants.py
@@ -1,0 +1,34 @@
+##################################################
+#               FOLDER NAMES                     #
+##################################################
+
+# Garmin Connect metrics - Folder names #
+_GARMIN_CONNECT_BASE_FOLDER = "garmin-connect"
+_GARMIN_CONNECT_BODY_COMPOSITION_FOLDER = (
+    _GARMIN_CONNECT_BASE_FOLDER + "-body-composition"
+)
+_GARMIN_CONNECT_HEART_RATE_FOLDER = _GARMIN_CONNECT_BASE_FOLDER + "-daily-heart-rate"
+_GARMIN_CONNECT_DAILY_SUMMARY_FOLDER = _GARMIN_CONNECT_BASE_FOLDER + "-daily-summary"
+_GARMIN_CONNECT_DAILY_PULSE_OX_FOLDER = _GARMIN_CONNECT_BASE_FOLDER + "-pulse-ox"
+_GARMIN_CONNECT_SLEEP_PULSE_OX_FOLDER = _GARMIN_CONNECT_BASE_FOLDER + "-sleep-pulse-ox"
+_GARMIN_CONNECT_DAILY_RESPIRATION_FOLDER = _GARMIN_CONNECT_BASE_FOLDER + "-respiration"
+_GARMIN_CONNECT_SLEEP_RESPIRATION_FOLDER = (
+    _GARMIN_CONNECT_BASE_FOLDER + "-sleep-respiration"
+)
+_GARMIN_CONNECT_SLEEP_STAGE_FOLDER = _GARMIN_CONNECT_BASE_FOLDER + "-sleep-stage"
+_GARMIN_CONNECT_SLEEP_SUMMARY_FOLDER = _GARMIN_CONNECT_BASE_FOLDER + "-sleep-summary"
+_GARMIN_CONNECT_EPOCH_FOLDER = _GARMIN_CONNECT_BASE_FOLDER + "-epoch"
+_GARMIN_CONNECT_STRESS_FOLDER = _GARMIN_CONNECT_BASE_FOLDER + "-stress"
+
+# Garmin Device metrics - Folder names #
+_GARMIN_DEVICE_FOLDER = "garmin-device"
+_GARMIN_DEVICE_BBI_FOLDER = _GARMIN_DEVICE_FOLDER + "-bbi"
+_GARMIN_DEVICE_HEART_RATE_FOLDER = _GARMIN_DEVICE_FOLDER + "-heart-rate"
+_GARMIN_DEVICE_PULSE_OX_FOLDER = _GARMIN_DEVICE_FOLDER + "-pulse-ox"
+_GARMIN_DEVICE_RESPIRATION_FOLDER = _GARMIN_DEVICE_FOLDER + "-respiration"
+_GARMIN_DEVICE_STEP_FOLDER = _GARMIN_DEVICE_FOLDER + "-step"
+_GARMIN_DEVICE_STRESS_FOLDER = _GARMIN_DEVICE_FOLDER + "-stress"
+
+# Questionnaire and Todo - Folder names #
+_QUESTIONNAIRE_FOLDER = "questionnaire"
+_TODO_FOLDER = "todo"
